requirements.txt
torch==1.13.1
torchvision==0.14.1
matplotlib==3.5.3
pandas==1.3.5
tensorboard==2.11.2
hydra-core==1.3.2